+++ mypy.ini
@@ -194,6 +194,9 @@ ignore_errors = True
 [mypy-torch.nn.parallel._functions]
 ignore_errors = True
 
+[mypy-torch.nn.parallel.comm]
+ignore_errors = True
+
 [mypy-torch.nn.quantized.functional]
 ignore_errors = True
